requirements.txt
requests==2.32.3
python-dateutil==2.9.0.post0
gspread==6.1.4
google-auth==2.36.0
cryptography==44.0.0
tzdata==2025.2
python-dotenv==1.0.1
urllib3==1.26.18

